+++ src/components/button/CloseButton.styled.ts
